fix(i18n): preserve mapped filters and partial pages
